--- go.mod
@@ -1,3 +1,0 @@
-module go-url-shortener
-
-go 1.21
