+++ url-shortener/Procfile
@@ -1,0 +1,1 @@
+web: exec bin/go-url-shortener
