+++ url-shortener/go.mod
@@ -1,0 +1,3 @@
+module go-url-shortener
+
+go 1.21
