+++ url-shortener/models/models.go
@@ -1,0 +1,12 @@
+package models
+
+// Struct to read the request payload
+type ShortenRequest struct {
+	URL string `json:"url"`
+	CustomUrl string `json:"custom_url"`
+}
+
+// Struct for the response payload
+type ShortenResponse struct {
+	ShortURL string `json:"short_url"`
+}
